update system to put in 'experiment' mode
new page was created 'preexperimento.ejs'. with this page admninstrator will collect info to validate the algorythm of the system responsible for generate the article to the end user. Also the info collected and the results accomplished can be seen as soon as the writing of thesis finish.
other changes like layout ajustment and new routes was added to make the contact between app and the database (register of the experiment data).